requirements.txt
pip==24.3.1
setuptools==75.6.0
wheel == 0.45.1
requests==2.32.3
pyyaml==6.0.2
numpy==2.2.0
xlrd==2.0.1
pandas==2.2.3
cloudscraper==1.2.71
pytest-httpserver==1.1.0
python-decouple==3.8
python-telegram-bot[job-queue]==21.9
bpython == 0.24
certifi==2024.12.14
garth==0.5.2